feat: add vite to the project
