+++ Client/pages/auth.html
@@ -1,0 +1,115 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <link rel="preconnect" href="https://fonts.googleapis.com" />
+    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
+    <link rel="preconnect" href="https://fonts.googleapis.com" />
+    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
+    <link
+      href="https://fonts.googleapis.com/css2?family=Great+Vibes&family=Merienda:wght@600&family=Poppins:ital,wght@0,300;0,900;1,300&family=Source+Serif+Pro:ital@1&display=swap"
+      rel="stylesheet"
+    />
+
+    <link
+      rel="stylesheet"
+      href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.3.0/css/all.min.css"
+      integrity="sha512-SzlrxWUlpfuzQ+pcUCosxcglQRNAq/DZjVsC0lE40xsADsfeQoEypE+enwcOiGjk/bSuGGKHEyjSoQ1zVisanQ=="
+      crossorigin="anonymous"
+      referrerpolicy="no-referrer"
+    />
+    <link rel="preconnect" href="https://fonts.googleapis.com" />
+    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
+    <link rel="preconnect" href="https://fonts.googleapis.com" />
+    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
+    <link
+      href="https://fonts.googleapis.com/css2?family=Poppins:ital,wght@0,300;0,900;1,300&family=Source+Serif+Pro:ital@1&display=swap"
+      rel="stylesheet"
+    />
+    <link
+      href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css"
+      rel="stylesheet"
+      integrity="sha384-QWTKZyjpPEjISv5WaRU9OFeRpok6YctnYmDr5pNlyT2bRjXh0JMhjY6hW+ALEwIH"
+      crossorigin="anonymous"
+    />
+    <link
+      rel="stylesheet"
+      href="https://cdn.jsdelivr.net/npm/swiper@11/swiper-bundle.min.css"
+    />
+    <script src="https://accounts.google.com/gsi/client" async defer></script>
+    <link rel="stylesheet" href="/Client/src/assets/css/auth.css" />
+    <title>E-commerce Authentification</title>
+  </head>
+  <body>
+    <header id="authHeader">
+      <h2>welcome to E-commerce</h2>
+      <p id="authMessage">
+        thanks for coming back, now get into your account my simple login
+        through your email and password.
+      </p>
+    </header>
+    <div class="login-auth" id="loginForm">
+      <label for="email"> Email Address</label>
+      <input
+        type="text"
+        id="email"
+        name="email"
+        placeholder="Your email address"
+      />
+      <div password-wrapper>
+      <label for="password">Password </label>
+      <div class="input-group">
+      <input
+        type="text"
+        id="password"
+        name="password"
+        placeholder="Must be at least 8 characters"
+        required
+        minlength="8"
+      />
+      <span class="toggle-password" id="togglePassword">
+        <i class="fas fa-eye-slash"></i>
+      </span>
+    </div>
+    </div>
+      <p>Forgot Password?<a href="reste_password.html">Reset Here</a></p>
+    </div>
+   
+
+     <!-- Add Google Sign-In section -->
+      <div class="social-login">
+        <div class="divider">
+          <span class="divider-line"></span>
+          <span class="divider-text">or continue with</span>
+          <span class="divider-line"></span>
+        </div>
+        <div class="google-signin">
+          <div id="googleSignInBtn" class="g-signin2" data-onsuccess="onGoogleSignIn"></div>
+          <!-- Custom styled button as fallback -->
+          <button id="customGoogleBtn" class="btn btn-outline-danger google-btn">
+            <i class="fab fa-google"></i> Sign in with Google
+          </button>
+        </div>
+      </div>
+    </div>
+    <div class="button-register">
+      <button id="loginBtn" type="button" class="btn btn-dark">join now</button>
+      <p>Don't have an account? <a href="register.html">Register Here</a></p>
+    </div>
+    <div id="userSection" style="display: none;">
+      <div class="user-info">
+        <img src="" alt="user Avatar" class="avatar" id="userAvatar">
+        <h3 id="userName"></h3>
+        <p id="userEmail"></p>
+      </div>
+      <button id="logoutBtn" class="btn btn-danger">Logout</button>
+    </div>
+
+    <!-- Add Google auth script -->
+    <script src="https://accounts.google.com/gsi/client" async defer></script>
+    <script src="/Client/src//assets/js/auth.js"></script>
+  </body>
+</html>
