+++ Client/pages/register.html
@@ -1,0 +1,124 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <link rel="preconnect" href="https://fonts.googleapis.com" />
+    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
+    <link rel="preconnect" href="https://fonts.googleapis.com" />
+    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
+    <link
+      href="https://fonts.googleapis.com/css2?family=Great+Vibes&family=Merienda:wght@600&family=Poppins:ital,wght@0,300;0,900;1,300&family=Source+Serif+Pro:ital@1&display=swap"
+      rel="stylesheet"
+    />
+
+    <link
+      rel="stylesheet"
+      href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.3.0/css/all.min.css"
+      integrity="sha512-SzlrxWUlpfuzQ+pcUCosxcglQRNAq/DZjVsC0lE40xsADsfeQoEypE+enwcOiGjk/bSuGGKHEyjSoQ1zVisanQ=="
+      crossorigin="anonymous"
+      referrerpolicy="no-referrer"
+    />
+    <link rel="preconnect" href="https://fonts.googleapis.com" />
+    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
+    <link rel="preconnect" href="https://fonts.googleapis.com" />
+    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
+    <link
+      href="https://fonts.googleapis.com/css2?family=Poppins:ital,wght@0,300;0,900;1,300&family=Source+Serif+Pro:ital@1&display=swap"
+      rel="stylesheet"
+    />
+    <link
+      href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css"
+      rel="stylesheet"
+      integrity="sha384-QWTKZyjpPEjISv5WaRU9OFeRpok6YctnYmDr5pNlyT2bRjXh0JMhjY6hW+ALEwIH"
+      crossorigin="anonymous"
+    />
+    <link
+      rel="stylesheet"
+      href="https://cdn.jsdelivr.net/npm/swiper@11/swiper-bundle.min.css"
+    />
+    <link rel="stylesheet" href="/Client/src/assets/css/register.css" />
+
+    <title>register</title>
+  </head>
+  <body>
+    <header>
+      <h2>welcome t, let's creat an account</h2>
+      <p>
+        create an account to keep track of your customers, and contributors.
+        Once your account has been created then don't forget to verify your
+        account through mail.
+      </p>
+    </header>
+    <div class="full_name">
+      <label for="firstName">First Name</label>
+      <input
+        type="text"
+        id="firstName"
+        name="firstName"
+        placeholder="Enter your first name"
+      />
+      <label for="LastName">Last Name</label>
+      <input
+        type="text"
+        id="LastName"
+        name="LastName"
+        placeholder="Enter your last name"
+      />
+    </div>
+    <div class="informations">
+      <label for="WorkEmail"> Work Email</label>
+      <input
+        type="text"
+        id="Email"
+        name="Email"
+        placeholder="Your work email"
+      />
+      <div class="input-group mb-3 select-container">
+        <label for="country">Country</label>
+        <select class="form-select" id="stateSelect" required name="country">
+          <option selected disabled>United States of America</option>
+        </select>
+      </div>
+
+      <div class="input-group mb-3">
+        <label for="inputGroupSelect01"
+          >Where did you first hear about page?</label
+        >
+        <select class="form-select" id="inputGroupSelect01" name="social">
+          <option selected>Select an option...</option>
+          <option>Facebook</option>
+          <option>Instagram</option>
+          <option>Youtube</option>
+          <option>Other...</option>
+        </select>
+      </div>
+    </div>
+
+    <label for="password">Password</label>
+    <input
+      type="password"
+      id="password"
+      name="password"
+      placeholder="Must be at least 8 characters."
+    />
+    <label for="confirmPassword">Confirm password</label>
+    <input
+      type="password"
+      id="confirmPassword"
+      name="confirmPassword"
+      placeholder="Confirm your password."
+    />
+    <p>
+      By contunuing you agree to the pige <a href="">terms of service</a> and
+      <a href="">privacy policy</a>
+    </p>
+    <div>
+      <button id="loginBtn" type="button" class="btn btn-dark">Register</button>
+      <p>Already registered? <a href="./auth.html">sing in</a></p>
+    </div>
+    <script src="/Client/src/assets/js/register.js"></script>
+  </body>
+</html>
