+++ Client/src/assets/css/auth.css
@@ -1,0 +1,238 @@
+* {
+  padding: 0;
+  margin: 0;
+  box-sizing: border-box;
+  list-style: none;
+  text-decoration: none !important;
+  font-family: "Poppins", sans-serif;
+}
+html {
+  overflow-x: hidden;
+}
+body {
+  overflow-x: hidden;
+}
+body {
+  margin: 0 auto 20px;
+  padding: 30px;
+  max-width: 500px;
+  background-color: #f8f9fa;
+  border-radius: 10px;
+  box-shadow: 0 2px 10px rgba(0, 0, 0, 0.1);
+}
+header {
+  text-align: center;
+  margin-bottom: 30px;
+}
+
+header h2 {
+  font-weight: 900;
+  margin-bottom: 10px;
+}
+
+.login-auth {
+  margin-bottom: 20px;
+}
+
+.login-auth label {
+  display: block;
+  text-align: left;
+  margin-bottom: 5px;
+  font-weight: bold;
+}
+.login-auth input {
+  width: 100%;
+  padding: 12px 0;
+  margin-bottom: 20px;
+  border: none;
+  border-bottom: 1px solid #ddd;
+  border-radius: 0;
+  background-color: transparent;
+  font-size: 15px;
+  transition: border-color 0.3s;
+}
+.login-auth input:focus {
+  transform: scale(1.01);
+  border-bottom: 2px solid #999;
+  box-shadow: 0 2px 5px rgba(0, 0, 0, 0.05);
+  outline: none;
+}
+.login-auth input::placeholder {
+  color: #999;
+  opacity: 1;
+}
+.login-auth input:hover {
+  transform: scale(1.01);
+  border-bottom: 2px solid #999;
+  box-shadow: 0 2px 5px rgba(0, 0, 0, 0.05);
+}
+
+/* help supprt password reset  */
+.support {
+  text-align: left !important;
+  margin: -10px 0 20px;
+}
+.login-auth p {
+  text-align: right;
+  margin: -10px 0 20px;
+}
+
+.login-auth a {
+  color: #0066cc;
+  text-decoration: none;
+  margin-left: 5px;
+}
+.password-wrapper {
+  position: relative;
+}
+.input-group {
+  position: relative;
+}
+.toggle-password {
+  position: absolute;
+  right: 10px;
+  top: 50%;
+  transform: translateY(-50%);
+  cursor: pointer;
+  color: #999;
+}
+.toggle-password:hover {
+  color: #666;
+}
+
+.button-register {
+  text-align: end;
+  margin-top: 20px;
+}
+
+.button-register button {
+  padding: 10px;
+  margin-bottom: 15px;
+  background-color: #000;
+  color: white;
+  border: none;
+  border-radius: 6px;
+  cursor: pointer;
+  transition: background-color 0.3s;
+}
+.button-register button:hover {
+  background-color: #333;
+}
+.button-register p {
+  color: #666;
+  font-size: 15px;
+}
+
+.button-register a {
+  color: #0066cc;
+  text-decoration: none;
+}
+/* Google Sign-In Styles */
+.social-login {
+  margin: 30px auto;
+  max-width: 450px;
+  padding: 0 30px;
+}
+
+.divider {
+  display: flex;
+  align-items: center;
+  margin: 20px 0;
+}
+
+.divider-line {
+  flex: 1;
+  height: 1px;
+  background-color: #ddd;
+}
+
+.divider-text {
+  padding: 0 10px;
+  color: #666;
+  font-size: 14px;
+}
+
+.google-signin {
+  display: flex;
+  flex-direction: column;
+  gap: 10px;
+}
+
+.google-btn {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  gap: 10px;
+  width: 100%;
+  padding: 10px;
+  border: 1px solid #ddd;
+  background: white;
+  color: #757575;
+  font-weight: 500;
+  transition: all 0.3s ease;
+}
+
+.google-btn:hover {
+  background: #f8f8f8;
+  color: #333;
+}
+
+/* Hide default Google button */
+.g-signin2 {
+  display: none;
+}
+/* Add these styles to your existing auth.css */
+#userSection {
+  text-align: center;
+  margin: 2rem auto;
+  max-width: 500px;
+  padding: 2rem;
+  border-radius: 10px;
+  background-color: #f8f9fa;
+  box-shadow: 0 2px 10px rgba(0, 0, 0, 0.1);
+}
+
+.user-info {
+  margin-bottom: 1.5rem;
+}
+
+.avatar {
+  width: 100px;
+  height: 100px;
+  border-radius: 50%;
+  object-fit: cover;
+  margin-bottom: 1rem;
+  display: none;
+}
+
+#logoutBtn {
+  margin-top: 1rem;
+  width: 70%;
+}
+
+.google-btn {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  gap: 8px;
+  width: 100%;
+  margin-top: 1rem;
+}
+
+.divider {
+  display: flex;
+  align-items: center;
+  margin: 1.5rem 0;
+}
+
+.divider-line {
+  flex: 1;
+  height: 1px;
+  background-color: #ddd;
+}
+
+.divider-text {
+  padding: 0 1rem;
+  color: #6c757d;
+  font-size: 0.9rem;
+}
