+++ Client/src/assets/css/styles.css
@@ -1,0 +1,1 @@
+@import "tailwindcss";
