+++ Client/src/assets/css/utility.css
@@ -1,0 +1,7 @@
+.center {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  text-align: center;
+}
