--- password.html
@@ -1,74 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="preconnect" href="https://fonts.googleapis.com" />
-    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
-    <link rel="preconnect" href="https://fonts.googleapis.com" />
-    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
-    <link
-      href="https://fonts.googleapis.com/css2?family=Great+Vibes&family=Merienda:wght@600&family=Poppins:ital,wght@0,300;0,900;1,300&family=Source+Serif+Pro:ital@1&display=swap"
-      rel="stylesheet"
-    />
-
-    <link
-      rel="stylesheet"
-      href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.3.0/css/all.min.css"
-      integrity="sha512-SzlrxWUlpfuzQ+pcUCosxcglQRNAq/DZjVsC0lE40xsADsfeQoEypE+enwcOiGjk/bSuGGKHEyjSoQ1zVisanQ=="
-      crossorigin="anonymous"
-      referrerpolicy="no-referrer"
-    />
-    <link rel="preconnect" href="https://fonts.googleapis.com" />
-    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
-    <link rel="preconnect" href="https://fonts.googleapis.com" />
-    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
-    <link
-      href="https://fonts.googleapis.com/css2?family=Poppins:ital,wght@0,300;0,900;1,300&family=Source+Serif+Pro:ital@1&display=swap"
-      rel="stylesheet"
-    />
-    <link
-      href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css"
-      rel="stylesheet"
-      integrity="sha384-QWTKZyjpPEjISv5WaRU9OFeRpok6YctnYmDr5pNlyT2bRjXh0JMhjY6hW+ALEwIH"
-      crossorigin="anonymous"
-    />
-    <link
-      rel="stylesheet"
-      href="https://cdn.jsdelivr.net/npm/swiper@11/swiper-bundle.min.css"
-    />
-    <link rel="stylesheet" href="auth.css" />
-    <title>reset password</title>
-  </head>
-  <body>
-    <header>
-      <h2>Create new password</h2>
-      <p>
-        Now create your new password and don't forget to add same password in
-        confirm password field regarding the use case.
-      </p>
-    </header>
-    <div class="login-auth">
-      <label for="email"> New password</label>
-      <input
-        type="text"
-        id="password"
-        name="password"
-        placeholder="Must be at least 8 characters."
-      />
-      <label for="password">Confirm Password</label>
-      <input
-        type="text"
-        id="password"
-        name="password"
-        placeholder="Must be at least 8 characters"
-        required
-        minlength="8"
-      />
-      <p>Need Help?<a href="">Contact Support</a></p>
-    </div>
-    <button type="button" class="btn btn-dark">Reset Password</button>
-  </body>
-</html>
