--- recover.css
@@ -1,19 +1,0 @@
-* {
-  padding: 0;
-  margin: 0;
-  box-sizing: border-box;
-  list-style: none;
-  text-decoration: none !important;
-  font-family: "Poppins", sans-serif;
-}
-.reset-email {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  gap: 1rem;
-}
-.reset-email p {
-  font-family: "Franklin Gothic Medium", "Arial Narrow", Arial, sans-serif;
-  color: gray;
-  font-size: 15px;
-}
