--- recover.html
@@ -1,29 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Password Recovery</title>
-    <link rel="stylesheet" href="auth.css" />
-    <link
-      rel="stylesheet"
-      href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.3.0/css/all.min.css"
-    />
-    <link rel="stylesheet" href="recover.css" />
-  </head>
-  <body>
-    <div class="reset-email">
-      <h2>Reset your password</h2>
-      <p>
-        we sent a password reset link to <span id="userEmail"></span> if there
-        was an account associated with it. Please allow up to 5 minutes for the
-        email to arrive.
-      </p>
-      <p>
-        if you are unable to access your account, please contact
-        <span><a href="">support@E-commer.com</a></span>
-      </p>
-    </div>
-    <script src="recover.js"></script>
-  </body>
-</html>
