+++ tailwind.config.js
@@ -1,0 +1,10 @@
+module.exports = {
+  content: [
+    "./index.html", // Adjust for any other file you want Tailwind to scan
+    "./src/**/*.{js,ts,jsx,tsx,html}",
+  ],
+  theme: {
+    extend: {},
+  },
+  plugins: [],
+};
